+++ Kokoro/docs/release-notes.md
@@ -10,6 +10,7 @@
 * For modders: Added resource cost action APIs.
 * For modders: Added continue/stop action APIs
 * For modders: Added an API to make outgoing X assignment actions.
+* For modders: Removed the `notnull` constraint from extension data.
 
 ## 1.1.1
 Released 28 December 2023.